requirements.txt
boto3
click
PyYaml
